changed to int

+++ setMotors.py
@@ -83,7 +83,7 @@ motor2B.start(0)
 if len(argv) <= 2:
     print("Need to call with x and y from commandline")
 else:
-    motorPWM = mixXY(float(argv[1]), float(argv[2]))
+    motorPWM = mixXY(int(argv[1]), int(argv[2]))
     leftMotorPWM = motorPWM[0]
     rightMotorPWM = motorPWM[1]
     print("left motor:",leftMotorPWM)
